pretty much done with most of the modules and errors and a little upgradation

+++ lib/firebase_resources/firestore_post.dart
@@ -1,13 +1,17 @@
 import 'dart:typed_data';
 
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:firebase_auth/firebase_auth.dart';
+import 'package:flutter/material.dart';
 
 import 'package:instaflurt3/firebase_resources/storage_method.dart';
 import 'package:instaflurt3/models/post.dart';
+import 'package:instaflurt3/utils/utils.dart';
 import 'package:uuid/uuid.dart';
 
 class FirestorePost {
   final FirebaseFirestore _firestore = FirebaseFirestore.instance;
+  final FirebaseAuth _auth = FirebaseAuth.instance;
 
   Future<String> uploadPost(
     String description,
@@ -20,9 +24,10 @@ class FirestorePost {
 
     try {
       String photoUrl =
-          await StorageMethod().uploadImageToStorage('Post Image', file, true);
+          await StorageMethod().uploadImageToStorage('Post_Image', file, true);
 
       String postId = const Uuid().v1();
+
       Post post = Post(
         description: description,
         profilePic: profilePic,
@@ -32,9 +37,10 @@ class FirestorePost {
         postId: postId,
         datePublished: DateTime.now(),
         postUrl: photoUrl,
-        
       );
+
       await _firestore.collection('posts').doc(postId).set(post.toJson());
+
       res = "success";
     } catch (err) {
       return err.toString();
@@ -50,6 +56,33 @@ class FirestorePost {
         });
       } else {
         await _firestore.collection('posts').doc(postId).update({
+          'likes': FieldValue.arrayUnion([uid])
+        });
+      }
+    } catch (e) {
+      print(e.toString());
+    }
+  }
+
+  Future<void> LikeComment(
+      String postId, String commentId, String uid, List likes) async {
+    try {
+      if (likes.contains(uid)) {
+        await _firestore
+            .collection('posts')
+            .doc(postId)
+            .collection('comments')
+            .doc(commentId)
+            .update({
+          'likes': FieldValue.arrayRemove([uid])
+        });
+      } else {
+        await _firestore
+            .collection('posts')
+            .doc(postId)
+            .collection('comments')
+            .doc(commentId)
+            .update({
           'likes': FieldValue.arrayUnion([uid])
         });
       }
@@ -88,7 +121,45 @@ class FirestorePost {
 
   Future<void> DeletePost(String postId) async {
     try {
-      await _firestore.collection('posts').doc(postId).delete();
+      DocumentSnapshot postSnap =
+          await _firestore.collection('posts').doc(postId).get();
+      if (postSnap.exists) {
+        if ((postSnap.data() as dynamic)['uid'] == _auth.currentUser!.uid) {
+          await _firestore.collection('posts').doc(postId).delete();
+        } else {
+          print("you cannot delete other user post");
+        }
+      } else {
+        print("Post does not exit");
+      }
+    } catch (e) {
+      print(e.toString());
+    }
+  }
+
+  Future<void> FollowUser(String uid, String followId) async {
+    try {
+      DocumentSnapshot snap =
+          await _firestore.collection('users').doc(uid).get();
+      List following = (snap.data()! as dynamic)['following'];
+
+      if (following.contains(followId)) {
+        await _firestore.collection('users').doc(followId).update({
+          'followers': FieldValue.arrayRemove([uid])
+        });
+
+        await _firestore.collection('users').doc(uid).update({
+          'following': FieldValue.arrayRemove([followId])
+        });
+      } else {
+        await _firestore.collection('users').doc(followId).update({
+          'followers': FieldValue.arrayUnion([uid])
+        });
+
+        await _firestore.collection('users').doc(uid).update({
+          'following': FieldValue.arrayUnion([followId])
+        });
+      }
     } catch (e) {
       print(e.toString());
     }
